+++ docs/pages/documentation/form/datepicker/examples/ExRange.vue
@@ -2,7 +2,6 @@
     <b-field label="Select a date">
         <b-datepicker
             placeholder="Click to select..."
-            :readonly="true"
             :min-date="minDate"
             :max-date="maxDate">
         </b-datepicker>